# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * A Woman With a Welder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductType = typeof ProductType[keyof typeof ProductType];


export const ProductType = {
  product: 'product',
  service: 'service',
} as const;

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  type: ProductType;
  available: boolean;
  createdAt: string;
}

export interface CartItem {
  productId: number;
  product: Product;
  quantity: number;
  shippingLabel?: string;
  shippingPrice?: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
  shippingLabel?: string;
  shippingPrice?: number;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  /** @nullable */
  memberId?: number | null;
  serviceId: number;
  status: BookingStatus;
  /** @nullable */
  preferredDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type BookingWithProductStatus = typeof BookingWithProductStatus[keyof typeof BookingWithProductStatus];


export const BookingWithProductStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingWithProduct {
  id: number;
  /** @nullable */
  memberId?: number | null;
  serviceId: number;
  status: BookingWithProductStatus;
  /** @nullable */
  preferredDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  service: Product;
}

export interface BookingInput {
  serviceId: number;
  preferredDate?: string;
  notes?: string;
  fullName?: string;
  address?: string;
  phone?: string;
  email?: string;
  serviceName?: string;
  servicePrice?: number;
  subtotal?: number;
  gst?: number;
  total?: number;
}

export interface Member {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface MemberRegisterInput {
  email: string;
  password: string;
  name: string;
}

export interface MemberLoginInput {
  email: string;
  password: string;
}

export interface ContactInput {
  name: string;
  email: string;
  message: string;
  phone?: string;
}

export type LogoutMember200 = {
  success: boolean;
};

export type SubmitContact200 = {
  success: boolean;
  message: string;
};