# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * A Woman With a Welder API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  serviceId: number;
  preferredDate?: string;
  notes?: string;
  fullName?: string;
  address?: string;
  phone?: string;
  email?: string;
  serviceName?: string;
  servicePrice?: number;
  subtotal?: number;
  gst?: number;
  total?: number;
}